feat: Implement welcome message encoding and handling
- Added a new WelcomeMessage struct and EncodeWelcomeMessage function in welcome.go for encoding connection IDs.
- Updated seed.go to use float32 for player coordinates.
- Refactored world_generator.go to use a switch statement for geometry types.
- Modified player.go to change coordinate types to float32 and added movement-related fields.
- Removed movement.go as it was no longer needed.
- Updated unit.go to change coordinate types to float32 and added player association.
- Created handle_message.go to centralize message handling for TCP and UDP connections.
- Enhanced socket.go to manage unique connection IDs and added UDP support.
- Implemented UDP message handling in udp_handler.go.
- Updated tcp_handler.go to send a welcome message upon connection.
- Adjusted message types in message_type.go to include WelcomeMessage.
- Updated unit_type.go to reorder unit types for clarity.

+++ migrations/seed.go
@@ -25,7 +25,7 @@ func Seed(db *gorm.DB) {
 	// player
 	testPlayers := []string{"Ryuzaki", "Muhammet", "Ahmet", "Mustafa"}
 	for _, player := range testPlayers {
-		testPlayer := models.Player{Nickname: player, CountryID: country.ID, CoordX: tile.CoordX, CoordY: tile.CoordY}
+		testPlayer := models.Player{Nickname: player, CountryID: country.ID, CoordX: float32(tile.CoordX), CoordY: float32(tile.CoordY)}
 		db.FirstOrCreate(&testPlayer, models.Player{Nickname: testPlayer.Nickname})
 	}
